server/requirements.txt
# Core server + API
Flask==3.1.1
flask-cors==6.0.1
gunicorn==23.0.0
python-dotenv==1.0.1  # ✅ Loads .env files

# AWS S3 support
boto3==1.39.3
botocore==1.39.3
s3transfer==0.13.0

# Image processing
pillow==11.3.0         # ✅ Use latest confirmed working with pillow-heif
pillow_heif==1.0.0     # ✅ Latest stable, supports HEIC via libheif

# Other required deps
requests==2.32.4
tqdm==4.67.1